requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
fastapi==0.103.2
geographiclib==2.0
geopy==2.4.0
h11==0.14.0
httptools==0.6.0
idna==3.4
iniconfig==2.0.0
mypy==1.6.0
mypy-extensions==1.0.0
packaging==23.2
pluggy==1.3.0
pydantic==2.4.2
pydantic_core==2.10.1
pytest==7.4.2
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
types-requests==2.31.0.8
types-ujson==5.8.0.1
typing_extensions==4.8.0
urllib3==2.0.6
uvicorn==0.23.2
watchfiles==0.20.0
websockets==11.0.3
